add service and fix code

+++ src/main/java/com/example/transactionpractice/service/RentService.java
@@ -1,12 +1,13 @@
 package com.example.transactionpractice.service;
 
-import com.example.transactionpractice.RentResponse;
+import com.example.transactionpractice.dto.RentResponse;
 
 public interface RentService {
 
-    public RentResponse getById(int id);
+    public RentResponse getById(long id);
 
-    public String addRent(int id);
+    public String addRent(long id);
 
-    public String rollbackRent(int id);
+    public String rollbackRent(long id);
+
 }
